engine: extract esm/layers/time_series_layer.js
